packaging: enforce new min. CloudStack version 4.15 starting GA/1.0 (#841)
There are many changes, including API changes in upstream master/4.15
which makes it challenging to maintain backward compability of Primate
with older versions of CloudStack. Therefore we need to ensure that the
rpm and deb Primate pkgs require CloudStack 4.15 as minimum version.
This would still leave some flexibility for advanced users of archive
builds (which adds risks that some features don't work with 4.14 or
older versions).

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>